Initialize items in work section
Creates a basic structure that has a predefined array of items and sets them with a key and an onClick function. The onClick calls the handleWorkFilter which will set the type of work to show later. In the className we check if the activeFilter is equal to the selected item and adds a new className called item-active.